+++ main.c
@@ -40,7 +40,7 @@ int main(int argc, char *argv[]) {
     struct file f = alloc_load_points_from_file(argv[1]);
     int dim = f.dimension;
     int lines = f.lines;
-    float *points = f.points;
+    double *points = f.points;
     int k = atoi(argv[2]);
     int n = lines;
     // int lda = n;
@@ -50,14 +50,14 @@ int main(int argc, char *argv[]) {
     //printf("Constructing fully connected matrix...\n");
     printf("Constructing the Laplacian ONE SHOT...\n");
     myInt64 start1 = start_tsc();
-    //float *fully_connected = calloc(lines * lines , sizeof(float));
+    //double *fully_connected = calloc(lines * lines , sizeof(double));
     //construct_fully_connected_matrix(points, lines, dim, fully_connected);
 
     printf("Constructing unnormalized Laplacian...\n");
     // compute unnormalized laplacian
-    //float *laplacian = malloc(lines * lines * sizeof(float));
+    //double *laplacian = malloc(lines * lines * sizeof(double));
     //construct_unnormalized_laplacian(fully_connected, lines, laplacian);
-    float *laplacian = calloc(lines*lines, sizeof(float));
+    double *laplacian = calloc(lines*lines, sizeof(double));
 
     if (dim >=8){
         oneshot_unnormalized_laplacian(points,lines,dim,laplacian);
@@ -67,7 +67,7 @@ int main(int argc, char *argv[]) {
     //compute the eigendecomposition and take the first k eigenvectors.
     myInt64 end1 = stop_tsc(start1);
     printf("Performing eigenvalue decomposition...\n");
-    // float *w = malloc(lines * sizeof(float));
+    // double *w = malloc(lines * sizeof(double));
     // lapack_int info = LAPACKE_dsyev(LAPACK_ROW_MAJOR, 'V', 'U', n, laplacian, lda, w);
     // /* Check for convergence */
     // if (info > 0) {
@@ -88,8 +88,8 @@ int main(int argc, char *argv[]) {
     //     }
     //     printf("\n");
     // }
-    float *eigenvalues = malloc(k * sizeof(float));
-    float *eigenvectors = malloc(n * k * sizeof(float));
+    double *eigenvalues = malloc(k * sizeof(double));
+    double *eigenvectors = malloc(n * k * sizeof(double));
     smallest_eigenvalues(laplacian, n, k, eigenvalues, eigenvectors);
 
     // print_matrix("Eigenvectors (stored columnwise)", n, n, laplacian, lda);
@@ -102,7 +102,7 @@ int main(int argc, char *argv[]) {
     // init datastructure
     struct cluster clusters[k];
     for (int i = 0; i < k; i++) {
-        clusters[i].mean = malloc(k * sizeof(float)); // k is the "dimension" here
+        clusters[i].mean = malloc(k * sizeof(double)); // k is the "dimension" here
         clusters[i].size = 0;
         clusters[i].indices = malloc(lines * sizeof(int)); // at most
     }
@@ -139,7 +139,7 @@ int main(int argc, char *argv[]) {
     // LEAVE THESE PRINTS (for the performance checking script)
     printf("%" PRIu64 "\n", runtime);
     printf("%" PRIu64 "\n", NUM_FLOPS);
-    printf("performance: %lf\n", (float)NUM_FLOPS/runtime);
+    printf("performance: %lf\n", (double)NUM_FLOPS/runtime);
 
 #ifdef VALIDATION
     char *my_argv; // = {"./base_clustering" , argv[1] , argv[2] , "./base_output"};
